[xcode] Merge avio_evbuffer into transcode
Only used by transcode anyway. Makes it easier to add custom seek for evbuffer.
